+++ src/oms_sensemaking/config.py
@@ -247,17 +247,17 @@ class MilSymbolSettings(BaseModel):
         description="Path to the rules config file"
     )
 
-class ObjectMinimumsSettings(BaseModel):
-    enable_object_minimums_sensemaker: bool = Field(
+class ObjectStandardsSettings(BaseModel):
+    enable_object_standards_sensemaker: bool = Field(
         True,
-        description="Toggle the object minimums sensemaker on/off"
+        description="Toggle the object standards sensemaker on/off"
     )
-    rmq_object_minimums_queue_name:str = Field(
-        "object-minimums-trigger",
-        description="the RMQ Object Minimums Queue name",
-        examples=["object-minimums-trigger"]
+    rmq_object_standards_queue_name:str = Field(
+        "object-standards-trigger",
+        description="the RMQ Object Standards Queue name",
+        examples=["object-standards-trigger"]
     )
-    rubrics_file_path: str = Field("./data/object_minimums.json", description="Path to the rubrics config file")
+    rubrics_file_path: str = Field("./data/object_standards.json", description="Path to the rubrics config file")
     max_rubric_hierarchy_levels: int = Field(
         5,
         description="Max number of class hierarchy levels to check when looking for a rubric (class + ancestors)",
@@ -522,7 +522,7 @@ class Settings(BaseSettings):
     )
 
     mil_symbol_settings: MilSymbolSettings = MilSymbolSettings()
-    object_minimum_settings: ObjectMinimumsSettings = ObjectMinimumsSettings()
+    object_standards_settings: ObjectStandardsSettings = ObjectStandardsSettings()
     incursion_settings: IncursionSettings = IncursionSettings()
     out_of_garrison_settings: OutOfGarrisonSettings = OutOfGarrisonSettings()
 
@@ -539,7 +539,7 @@ class Settings(BaseSettings):
     oms_crud_ttl_cache_size: int = Field(1024, description="Max items in ATOMS CRUD Tool's given TTL Cache")
     oms_crud_ttl_cache_seconds: int = Field(3600, description="Max time to live in ATOMS CRUD Tool's given TTL Cache")
     omsb_url: str = Field("https://graphql:8443/graphql", description="URL for OMSB")
-    omsb_version: str = Field("3.1.5", description="OMSB Version")
+    omsb_version: str = Field("3.1.6", description="OMSB Version")
     aac_url: str = Field("http://aac2:3000", description="URL for AAC")
     user_dn: str = Field(description="User DN")
     aac_cacert_path: str | None = Field(