+++ src/oms_sensemaking/core/events.py
@@ -318,7 +318,9 @@ class RabbitMQListener(BaseRabbitMQListener):
         try:
             audit_log: AuditLogEvent = AuditLogEvent.from_json(body.decode("utf-8"))
             audit_log.headers = HeaderParser().parse(properties)
-            LOGGER.info("%s Received %s %s: %s", self._name, audit_log.action, audit_log.objectType, audit_log.objectId)
+            LOGGER.debug(
+                "%s Received %s %s: %s", self._name, audit_log.action, audit_log.objectType, audit_log.objectId
+            )
             object_id = audit_log.objectId
 
             if self._event_filter and not self._event_filter.passes_filter(audit_log):
@@ -333,7 +335,7 @@ class RabbitMQListener(BaseRabbitMQListener):
                 return
 
             if self.handle_event and self.handle_event(audit_log):
-                LOGGER.info("Acknowledging processed object %s from %s", audit_log.objectId, self._queue_name)
+                LOGGER.debug("Acknowledging processed object %s from %s", audit_log.objectId, self._queue_name)
                 self._connection.add_callback_threadsafe(lambda: ch.basic_ack(delivery_tag=method.delivery_tag))
 
                 # Record successful processing metrics